Rewrite categories saga to TypeScript
