+++ src/store/categories/category.saga.ts
@@ -1,0 +1,30 @@
+import { takeLatest, call, put, all } from "typed-redux-saga";
+
+import { getCategoriesAndDocuments } from "../../utils/firebase/firebase.utils";
+
+import {
+  fetchCategoriesSuccess,
+  fetchCategoriesFailure,
+} from "./category.action";
+
+import { CATEGORIES_ACTION_TYPES } from "./category.types";
+
+export function* fetchCategoriesAsync() {
+  try {
+    const categoriesArray = yield* call(getCategoriesAndDocuments);
+    yield* put(fetchCategoriesSuccess(categoriesArray));
+  } catch (error) {
+    yield* put(fetchCategoriesFailure(error as Error));
+  }
+}
+
+export function* onFetchCategories() {
+  yield* takeLatest(
+    CATEGORIES_ACTION_TYPES.FETCH_CATEGORIES_START,
+    fetchCategoriesAsync
+  );
+}
+
+export function* categoriesSaga() {
+  yield* all([call(onFetchCategories)]);
+}
